\name{hier.part.data}
\alias{urban}
\alias{urban1}
\alias{chevan}
\title{
  Example data for hier.part
}
\description{
  Example data sets for hier.part package.
  }
\usage{
  data(urban)
  data(urban1)
  data(chevan)
}
\details{
    urban.txt

    Mean chlorophyll data (chl) for at 16 sites on 16 streams, with four
    catchment variables for each site (fimp, fconn, sep, unseal)
  
    urban1.txt

    Presence-absence data (pa) for a species in 58 sites, with four
    catchment variables as above

    chevan.txt

    Chi-squared (chisq) and R-squared (rsqu) for logistic and linear
    regression example from Chevan and Sutherland (1991)
}
\references{
  Chevan, A. and Sutherland, M. (1991) Hierarchical Partitioning. The
  American Statistician 45: 90-96  
}
\author{
  Chris Walsh \email{Chris.Walsh@sci.monash.edu.au}
}
\keyword{datasets}







