\name{hier.part.data}
\alias{urbanwq}
\alias{amphipod}
\alias{chevan}
\title{
  Example Data for hier.part package
}
\description{
  Example data sets for hier.part package.
  }
\usage{
  data(urbanwq)
  data(amphipod)
  data(chevan)
}
\details{
    urbanwq.txt

    Seven catchment variables (fimp, catchment imperviousness\eqn{^{0.25}};
    sconn, drainage connection\eqn{^{0.5}}; sdensep, septic tank
    density\eqn{^{0.5}}; unsealden, unsealed road density; fcarea,
    catchment area\eqn{^{0.25}}; selev, elevation\eqn{^{0.5}}; amgeast,
    longitude) and median baseflow concentrations of three water quality
    variables in sixteen streams draining 16 independent subcatchments
    (ldoc, log(dissolved organic carbon); lec, log(electrical
    conductivity); lnox, log(nitrate/nitrite). Data from Hatt et
    al. (2004).

    amphipod.txt

    Presence-absence data for a stream-dwelling amphipod (Austrogammarus
    australis) in 58 sites, with four catchment variables (fimp,
    imperviousness\eqn{^{0.25}};fconn, drainage connection\eqn{^{0.25}};
    densep, density of septic tanks; unseal, unsealed road density).
    Data from Walsh et al. (2004).

    chevan.txt

    Chi-squared (chisq) and R-squared (rsqu) for logistic and linear
    regression example from Chevan and Sutherland (1991).
}
\references{
  Chevan, A. and Sutherland, M. 1991 Hierarchical Partitioning. \emph{The
  American Statistician} \bold{45}, 90--96.

  Hatt, B. E., Fletcher, T. D., Walsh, C. J. and Taylor, S. L. 2004 The
  influence of urban density and drainage infrastructure on the
  concentrations and loads of pollutants in small
  streams. \emph{Environmental Management} \bold{34}, 112--124.

  Walsh, C. J., Papas, P. J., Crowther, D., Sim, P. T., and Yoo, J. 2004
  Stormwater drainage pipes as a threat to a stream-dwelling amphipod of
  conservation significance, \emph{Austrogammarus australis}, in southeastern
  Australia. \emph{Biodiversity and Conservation} \bold{13}, 781--793.

}
\author{
  Chris Walsh \email{cwalsh@unimelb.edu.au}
}

\keyword{datasets}




